ENA-3950
- major refactor of normalization and spot data reading code.
