+++ src/main/java/uk/ac/ebi/ena/readtools/loader/fastq/FastqIterativeConsumer.java
@@ -10,10 +10,10 @@
 */
 package uk.ac.ebi.ena.readtools.loader.fastq;
 
+import uk.ac.ebi.ena.readtools.common.reads.QualityNormalizer;
+
 import java.io.File;
 import java.util.Iterator;
-
-import uk.ac.ebi.ena.readtools.loader.common.QualityNormalizer;
 
 public class
 FastqIterativeConsumer implements Iterable<FastqSpot>
