+++ src/main/java/uk/ac/ebi/ena/readtools/loader/fastq/PairedFastqConsumer.java
@@ -87,7 +87,7 @@ PairedFastqConsumer extends AbstractPagedDataConsumer<DataSpot, FastqSpot>
             readIndexStr = getReadIndex( spot.bname );
         } catch ( DataConsumerException de )
         {
-            readIndexStr = spot.getReadIndex();
+            readIndexStr = spot.readIndex;
         }
 
         int readIndex = Integer.parseInt( readIndexStr ) - 1;
